Optimized code

+++ app/src/main/java/com/ljyh/music/data/model/qq/u/SearchResult.kt
@@ -181,7 +181,7 @@ data class SearchResult(
                             @SerializedName("alert")
                             val alert: Int,
                             @SerializedName("icon2")
-                            val icon2: Int,
+                            val icon2: Long,
                             @SerializedName("icons")
                             val icons: Int,
                             @SerializedName("msgdown")
